kinesis: move thumb alt & ctrl modifiers
I would like to experiment with making my secondary left, large thumb
key to be a LT. The first step in this is moving CTRL to a different
location. I moved it to the END position as I tried the LALT position,
but typing common vim comamnds (C-u, C-e, C-o) were too difficult.

During this change, I moved LALT to my right thumb cluster so that I
could actually perform Opt-BS, which I use often. This replaces the
need for the OSL-D command to delete back a word.