kinesis: home-row shift; reimplement shifted keycodes
Adds a shift modifier as a home-row modifier on each middle finger. I
was getting a lot of ulnar nerve strain and it seems to flare up because
of the need to twist my hard to hold down shift. The home-row mods were
implemented according to this guide:
https://precondition.github.io/home-row-mods#using-home-row-mods-with-qmk

There was one (somewhat major) problem with home-row mods that I needed
to work around that caused me to re-implement my shifted keycodes.
Because my shifted keycodes for the number row (and other various keys)
are not standard, pressing the home-row shift + KC_PLUS for example
didn't produce KC_1 as I needed. This is because my shifted keycodes
were implemented in a separate layer. I could not use this alternate
layer in my home-row mods because MT() doesn't support any non-modifier
keys (see:
https://beta.docs.qmk.fm/using-qmk/advanced-keycodes/mod_tap).

To work around this I overrode the meaning of each of the keys with
non-default shifted keycodes inside process_record_user to send the
correct keycode when shift + that keycode were registered. I found this
idea here:
https://www.reddit.com/r/olkb/comments/k8wrge/which_layer_is_active_when_shift_key_is_held_down/gf1pb43?utm_source=share&utm_medium=web2x&context=3). Overall I like this solution much better than the separate shifted layer anyway as I don't need to keep muliple layers and layouts in sync with updated symbol positions. This change also allows for shift to be registered when doing complex chords (i.e. Alt+Ctrl+Shift+A) which was a problem I had to work around before when using a custom coded layer. Even if I don't stick with home-row mods long term I'm definitely going to keep this implementation of shifted keycodes.